+++ sdim/__init__.py
@@ -66,7 +66,7 @@ Measured qudit (3) as (1) and was deterministic
 
 from .circuit_io import read_circuit, write_circuit, circuit_to_cirq_circuit, cirq_statevector_from_circuit
 from .program import Program
-from .random_circuit import generate_random_circuit, generate_and_write_random_circuit
+from .random_circuit import generate_random_clifford_circuit, generate_and_write_random_circuit
 from .circuit import Circuit
 from .tableau.tableau_composite import WeylTableau
 from .tableau.dataclasses import MeasurementResult, Tableau