+++ sdim/tableau/dataclasses.py
@@ -5,46 +5,13 @@ import numpy as np
 from math import gcd
 
 @dataclass
-class MeasurementResult:
-    """
-    Represents the result of a qudit measurement in a quantum circuit.
-
-    Attributes:
-        qudit_index (int): The index of the measured qudit.
-        deterministic (bool): Whether the measurement result was deterministic.
-        measurement_value (int): The measured value of the qudit.
-    """
-
-    qudit_index: int
-    deterministic: bool
-    measurement_value: int
-
-    def __str__(self) -> str:
-        """
-        Returns a string representation of the measurement result.
-
-        Returns:
-            str: A human-readable description of the measurement result.
-        """
-        measurement_type_str = "deterministic" if self.deterministic else "random"
-        return f"Measured qudit ({self.qudit_index}) as ({self.measurement_value}) and was {measurement_type_str}"
-
-    def __repr__(self) -> str:
-        """
-        Returns a string representation of the measurement result.
-
-        Returns:
-            str: Same as __str__ method.
-        """
-        return str(self)
-
-@dataclass
 class Tableau:
     """
     Represents a stabilizer tableau for a quantum circuit simulation.
 
     This class encapsulates the phase vector, Z block, and X block that 
     describe the state of a quantum system in the stabilizer formalism.
+    The stabilizers are stored as columns in the Z and X blocks.
 
     Attributes:
         num_qudits (int): The number of qudits in the system.
@@ -71,6 +38,13 @@ class Tableau:
         if self.x_block is None:
             self.x_block = np.zeros((self.num_qudits, self.num_qudits), dtype=np.int64)
 
+    def modulo(self):
+        """
+        Applies the modulo operators to the phase vector and stabilizers according to the order and dimension
+        """
+        self.z_block %= self.dimension
+        self.x_block %= self.dimension
+        self.phase_vector %= self.order
     @cached_property
     def coprime_order(self) -> set:
         """
@@ -187,3 +161,50 @@ class Tableau:
         self.print_phase_vector()
         self.print_z_block()
         self.print_x_block()
+
+
+@dataclass
+class MeasurementResult:
+    """
+    Represents the result of a qudit measurement in a quantum circuit.
+
+    Attributes:
+        qudit_index (int): The index of the measured qudit.
+        deterministic (bool): Whether the measurement result was deterministic.
+        measurement_value (int): The measured value of the qudit.
+    """
+
+    qudit_index: int
+    deterministic: bool
+    measurement_value: int
+    stabilizer_tableau: Optional[Tableau] = None
+    
+    def __str__(self) -> str:
+        """
+        Returns a string representation of the measurement result.
+
+        Returns:
+            str: A human-readable description of the measurement result.
+        """
+        measurement_type_str = "deterministic" if self.deterministic else "random"
+        return f"Measured qudit ({self.qudit_index}) as ({self.measurement_value}) and was {measurement_type_str}"
+
+    def __repr__(self) -> str:
+        """
+        Returns a string representation of the measurement result.
+
+        Returns:
+            str: Same as __str__ method.
+        """
+        return str(self)
+    
+    def get_tableau(self):
+        """
+        Returns the stabilizer tableau.
+
+        Returns:
+            Tableau: The stabilizer tableau.
+        """
+        if self.stabilizer_tableau is None:
+            raise ValueError("Stabilizer tableau not recorded during measurement")
+        return self.stabilizer_tableau