+++ sdim/tableau/tableau_gates.py
@@ -7,9 +7,10 @@ from .tableau_composite import WeylTableau
 from .tableau_prime import ExtendedTableau
 from .dataclasses import MeasurementResult, Tableau
 import numpy as np
+import random
     
 # Gate application functions
-def apply_I(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_I(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply identity gate (do nothing).
 
@@ -23,7 +24,7 @@ def apply_I(tableau: Tableau, qudit_index: int, _) -> None:
     """
     return None
 
-def apply_X(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_X(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply Pauli X gate.
 
@@ -52,7 +53,7 @@ def apply_X(tableau: Tableau, qudit_index: int, _) -> None:
             tableau.hadamard(qudit_index)
     return None
 
-def apply_X_inv(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_X_inv(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply Pauli X inverse gate.
 
@@ -81,7 +82,7 @@ def apply_X_inv(tableau: Tableau, qudit_index: int, _) -> None:
             tableau.hadamard(qudit_index)
     return None
 
-def apply_Z(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_Z(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply Pauli Z gate.
 
@@ -108,7 +109,7 @@ def apply_Z(tableau: Tableau, qudit_index: int, _) -> None:
             tableau.hadamard(qudit_index)
     return None
 
-def apply_Z_inv(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_Z_inv(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply Pauli Z inverse gate.
 
@@ -135,7 +136,7 @@ def apply_Z_inv(tableau: Tableau, qudit_index: int, _) -> None:
             tableau.hadamard_inv(qudit_index)
     return None
 
-def apply_H(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_H(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply Hadamard gate.
 
@@ -150,7 +151,7 @@ def apply_H(tableau: Tableau, qudit_index: int, _) -> None:
     tableau.hadamard(qudit_index)
     return None
 
-def apply_H_inv(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_H_inv(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply Hadamard inverse gate.
 
@@ -165,7 +166,7 @@ def apply_H_inv(tableau: Tableau, qudit_index: int, _) -> None:
     tableau.hadamard_inv(qudit_index)
     return None
 
-def apply_P(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_P(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply Phase gate.
 
@@ -180,7 +181,7 @@ def apply_P(tableau: Tableau, qudit_index: int, _) -> None:
     tableau.phase(qudit_index)
     return None
 
-def apply_P_inv(tableau: Tableau, qudit_index: int, _) -> None:
+def apply_P_inv(tableau: Tableau, qudit_index: int, *_) -> None:
     """
     Apply Phase inverse gate.
 
@@ -195,7 +196,7 @@ def apply_P_inv(tableau: Tableau, qudit_index: int, _) -> None:
     tableau.phase_inv(qudit_index)
     return None
 
-def apply_CNOT(tableau: Tableau, control: int, target: int) -> None:
+def apply_CNOT(tableau: Tableau, control: int, target: int, *_) -> None:
     """
     Apply CNOT gate.
 
@@ -210,7 +211,7 @@ def apply_CNOT(tableau: Tableau, control: int, target: int) -> None:
     tableau.cnot(control, target)
     return None
 
-def apply_CNOT_inv(tableau: Tableau, control: int, target: int) -> None:
+def apply_CNOT_inv(tableau: Tableau, control: int, target: int, *_) -> None:
     """
     Apply CNOT inverse gate.
 
@@ -225,7 +226,7 @@ def apply_CNOT_inv(tableau: Tableau, control: int, target: int) -> None:
     tableau.cnot_inv(control, target)
     return None
 
-def apply_CZ(tableau: Tableau, control: int, target: int) -> None:
+def apply_CZ(tableau: Tableau, control: int, target: int, *_) -> None:
     """
     Apply CZ gate.
 
@@ -237,12 +238,12 @@ def apply_CZ(tableau: Tableau, control: int, target: int) -> None:
     Returns:
         None
     """
-    tableau.hadamard(target)
-    tableau.cnot(control, target)
     tableau.hadamard_inv(target)
+    tableau.cnot(control, target)
+    tableau.hadamard(target)
     return None
 
-def apply_CZ_inv(tableau: Tableau, control: int, target: int) -> None:
+def apply_CZ_inv(tableau: Tableau, control: int, target: int, *_) -> None:
     """
     Apply CZ inverse gate.
 
@@ -254,12 +255,12 @@ def apply_CZ_inv(tableau: Tableau, control: int, target: int) -> None:
     Returns:
         None
     """
-    tableau.hadamard(target)
-    tableau.cnot_inv(control, target)
     tableau.hadamard_inv(target)
+    tableau.cnot_inv(control, target)
+    tableau.hadamard(target)
     return None
 
-def apply_measure(tableau: Tableau, qudit_index: int, _) -> Optional[MeasurementResult]:
+def apply_measure(tableau: Tableau, qudit_index: int, *_) -> Optional[MeasurementResult]:
     """
     Apply measurement.
 
@@ -276,7 +277,7 @@ def apply_measure(tableau: Tableau, qudit_index: int, _) -> Optional[Measurement
     else:
         return tableau.measure(qudit_index)
 
-def apply_measure_x(tableau: Tableau, qudit_index: int, _) -> Optional[MeasurementResult]:
+def apply_measure_x(tableau: Tableau, qudit_index: int, *_) -> Optional[MeasurementResult]:
     """
     Apply X-basis measurement.
 
@@ -288,17 +289,18 @@ def apply_measure_x(tableau: Tableau, qudit_index: int, _) -> Optional[Measureme
     Returns:
         Optional[MeasurementResult]: The result of the measurement, if applicable.
     """
-    tableau.hadamard(qudit_index)
+    tableau.hadamard_inv(qudit_index)
     if isinstance(tableau, WeylTableau):
         return tableau.measure_z(qudit_index)
     else:
         return tableau.measure(qudit_index)
 
-def apply_SWAP(tableau: Tableau, qudit_index: int, target_index: int) -> None:
+def apply_SWAP(tableau: Tableau, qudit_index: int, target_index: int, *_) -> None:
     """
     Apply SWAP gate.
 
-    Taken from Beaudrap Lemma 6 (eq 19)
+    Taken from Beaudrap Lemma 6 (eq 19) for composite case
+    Taken from Farinholt Figure 1 for prime case.
 
     Args:
         tableau (Tableau): The quantum tableau.
@@ -308,16 +310,27 @@ def apply_SWAP(tableau: Tableau, qudit_index: int, target_index: int) -> None:
     Returns:
         None
     """
-    tableau.cnot(qudit_index, target_index)
-    tableau.cnot_inv(target_index, qudit_index)
-    tableau.cnot(qudit_index, target_index)
-    tableau.hadamard(qudit_index)
-    tableau.hadamard(qudit_index)
+    if isinstance(tableau, WeylTableau):
+        tableau.cnot(qudit_index, target_index)
+        tableau.cnot_inv(target_index, qudit_index)
+        tableau.cnot(qudit_index, target_index)
+        tableau.hadamard(qudit_index)
+        tableau.hadamard(qudit_index)
+    else:
+        tableau.cnot(qudit_index, target_index)
+        tableau.hadamard(qudit_index)
+        tableau.hadamard(target_index)
+        tableau.cnot(qudit_index, target_index)
+        tableau.hadamard(qudit_index)
+        tableau.hadamard(target_index)
+        tableau.cnot(qudit_index, target_index)
+        tableau.hadamard(target_index)
+        tableau.hadamard(target_index)
     return None
 
-def apply_reset(tableau: Tableau, qudit_index: int, _) -> Optional[MeasurementResult]:
+def apply_reset(tableau: Tableau, qudit_index: int, *_) -> Optional[MeasurementResult]:
     """
-    Apply reset gate. IF the qudit is measured to be in the $\ket{1}$ state, reset it to the $\ket{0}$ state.
+    Apply reset gate. IF the qudit is measured to be in the $/ket{1}$ state, reset it to the $/ket{0}$ state.
 
     Args:
         tableau (Tableau): The quantum tableau.
@@ -331,4 +344,73 @@ def apply_reset(tableau: Tableau, qudit_index: int, _) -> Optional[MeasurementRe
         return tableau.measure_z(qudit_index)
     else:
         return tableau.measure(qudit_index)
+    
+
+def apply_single_qudit_noise(tableau : Tableau, qudit_index: int, _, params: dict) -> None:
+    """
+    DEPRECIATED -- Noise behavior is handled in program.py in Pauli frames.  This gate definition is nonetheless included for the sake of documentation and completeness.
+    
+    Given a probability and noise channel type, probabilistically applies a qudit Pauli to a target qudit.  Supports channels for flips, phase dampening, and depolarizing noise.  
+    Defaults to applying depolarizing noise with probability 0.5 if no parameters are passed.  
+    For a tableau of dimension d and with a probability argument p, the channels behave in the following way:
+
+    Flip: Applies a qudit Pauli of the form X^a on the target qudit with probability p, where a is randomly sampled from {0, 1, ..., d - 1}.
+
+    Phase dampening: Applies a qudit Pauli of the form Z^a on the target qudit with probability p, where a is randomly sampled from {0, 1, ..., d - 1}.
+
+    Depolarizing: Applies a qudit Pauli of the form of X^a Z^b with probability p, where a and b are elements of {0, 1, ..., d - 1}, and the tuple (a, b) is randomly sampled from the set {(a, b) | (a != 0) OR (b != 0)}.
+
+    Args:
+        tableau (Tableau): The quantum tableau.
+        qudit_index (int): The index of the qudit on which to apply noise.
+        _ : Unused target index.
+        noise_channel ((named) str): The type of noise applied to the target qudit.  Flip, phase, and depolarizing errors are "f", "p", and "d", respectively.
+        prob ((named) float): The probability that a non-identity Pauli gate is applied to the target qudit.
+
+    Returns:
+        None  
+    """
+    prob = float(params.get('prob', 0.5))
+    noise_channel = params.get('noise_channel', 'd')
+
+    if noise_channel == 'd':
+        num = random.uniform(0.0, 1.0)
+        # Determine whether to apply I or not
+        if num < 1.0 - prob:
+            return None
+        
+        # Sample error from all non-identity Pauli gates 
+        elem = random.randint(1, tableau.dimension - 1)
+        x_exp = elem % tableau.dimension
+        z_exp = elem // tableau.dimension
+
+        # Apply the Pauli gates.  Order does not matter up to phase since measurement statistics are identical.
+        for _ in range(x_exp):
+            apply_X(tableau=tableau, qudit_index=qudit_index)
+        for _ in range(z_exp):
+            apply_Z(tableau=tableau, qudit_index=qudit_index)
+
+        return None
+    
+    elif noise_channel == 'f' or noise_channel == 'p':
+        num = random.uniform(0.0, 1.0)
+        flip = (noise_channel == 'f')
+        # Determine whether to apply I or not
+        if num < 1.0 - prob:
+            return None
+        # Sample non-identity Pauli exponent
+        op_exp = random.randint(1, tableau.dimension - 1)
+        # Apply appropriate error
+        if flip:
+            for _ in range(op_exp):
+                apply_X(tableau=tableau, qudit_index=qudit_index)
+        else:
+            for _ in range(op_exp):
+                apply_Z(tableau=tableau, qudit_index=qudit_index)
+        
+        return None
+    
+
+    raise ValueError("Must specify a valid noise channel.")
+